--fix JS import works
We had to fix the `apply` method, which didn't return the unmodified content when there was no plugin import block (so there was nothing to change).

Now replaceImports doesn't get the raw content, we are only collecting the replaceStrings from it, and doing the replacement in the apply block, once everything has been loaded.