--enh Handle sass-extract's own importer call separately
While `node-sass` expects `done()` to call, when ready, `sass-extract` needs a `Promise` to return.

We need to provide both possibilities even for `sass-extract`, as it makes the extraction job in two rounds – in first step it passes over the importer to node-sas (which takes the callback-version), and in the second step, sass-extract itself needs the promised version.